Move Android file methods into separate class and update tests
