--- spec/test-data/test.xcconfig
@@ -1,3 +1,0 @@
-BUILD_NUMBER = 1234
-VERSION_LONG = 2023.19.0.0
-VERSION_SHORT = 2023.19
